Template scalar assembler over float type.
