+++ cpp/dolfinx/fem/CMakeLists.txt
@@ -23,7 +23,6 @@ set(HEADERS_fem
 target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/assembler.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_matrix_impl.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/assemble_scalar_impl.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/assemble_vector_impl.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/CoordinateElement.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/DirichletBC.cpp
